requirements.txt
Django==4.2
psycopg2-binary
gunicorn
whitenoise
djangorestframework>=3.14